% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilVAE-package.R
\docType{package}
\name{soilVAE-package}
\alias{soilVAE}
\alias{soilVAE-package}
\title{soilVAE: Supervised Variational Autoencoder Regression (Keras/TensorFlow via reticulate)}
\description{
The soilVAE package provides a supervised Variational Autoencoder (VAE) regression model
and tuning utilities for high-dimensional predictors (e.g., soil spectroscopy).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hugomachadorodrigues.github.io/soilVAE}
  \item \url{https://github.com/HugoMachadoRodrigues/soilVAE}
  \item Report bugs at \url{https://github.com/HugoMachadoRodrigues/soilVAE/issues}
}

}
\author{
\strong{Maintainer}: Hugo Rodrigues \email{rodrigues.machado.hugo@gmail.com}

}
\keyword{internal}
