% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\docType{package}
\name{soilVAE}
\alias{soilVAE-package}
\alias{soilVAE}
\title{soilVAE: Supervised VAE regression via reticulate}
\description{
Tools to build, tune, and use a supervised variational autoencoder (VAE) regression model for soil spectroscopy and related high-dimensional predictors. The model is implemented in Python (Keras 3 / TensorFlow) and accessed from R via reticulate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YOUR_GITHUB/soilVAE}
  \item Report bugs at \url{https://github.com/YOUR_GITHUB/soilVAE/issues}
}

}
\author{
\strong{Maintainer}: Hugo Rodrigues \email{rodrigues.machado.hugo@gmail.com}

}
