% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R, R/vae_build.R
\name{vae_build}
\alias{vae_build}
\title{Build a supervised VAE regression model (VAEReg)}
\usage{
vae_build(
  input_dim,
  hidden_enc = c(512L, 256L),
  hidden_dec = c(256L, 512L),
  latent_dim = 32L,
  dropout = 0.1,
  lr = 0.001,
  beta_kl = 1,
  alpha_y = 1
)

vae_build(
  input_dim,
  hidden_enc = c(512L, 256L),
  hidden_dec = c(256L, 512L),
  latent_dim = 32L,
  dropout = 0.1,
  lr = 0.001,
  beta_kl = 1,
  alpha_y = 1
)
}
\arguments{
\item{input_dim}{integer}

\item{hidden_enc}{integer vector}

\item{hidden_dec}{integer vector}

\item{latent_dim}{integer}

\item{dropout}{numeric}

\item{lr}{numeric learning rate}

\item{beta_kl}{numeric}

\item{alpha_y}{numeric}
}
\value{
Python keras model object (VAEReg)
}
\description{
Build a supervised VAE regression model (VAEReg)

Build a supervised VAE regression model (VAEReg)
}
