% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R, R/py_setup.R, R/vae_configure.R
\name{vae_configure}
\alias{vae_configure}
\title{Configure Python / reticulate for soilVAE}
\usage{
vae_configure(python = NULL, venv = NULL, conda = NULL)

vae_configure(python = NULL, venv = NULL, conda = NULL)

vae_configure(python = NULL, venv = NULL, conda = NULL)
}
\arguments{
\item{python}{Path to python executable (optional).}

\item{venv}{Name of virtualenv to use (optional).}

\item{conda}{Name of conda env to use (optional).}
}
\value{
TRUE invisibly.
}
\description{
Configure Python / reticulate for soilVAE

Configure Python / reticulate for soilVAE

Configure Python / reticulate for soilVAE
}
