% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R, R/vae_predict.R
\name{vae_encode}
\alias{vae_encode}
\title{Extract latent embeddings (z) from VAEReg}
\usage{
vae_encode(model, X)

vae_encode(model, X)
}
\arguments{
\item{model}{Python VAEReg}

\item{X}{matrix}
}
\value{
matrix
}
\description{
Extract latent embeddings (z) from VAEReg

Extract latent embeddings (z) from VAEReg
}
