% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R, R/vae_predict.R
\name{vae_predict}
\alias{vae_predict}
\title{Predict y using VAEReg (via latent z -> y_head)}
\usage{
vae_predict(model, X)

vae_predict(model, X)
}
\arguments{
\item{model}{Python VAEReg}

\item{X}{matrix}
}
\value{
numeric vector
}
\description{
Predict y using VAEReg (via latent z -> y_head)

Predict y using VAEReg (via latent z -> y_head)
}
