% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{select_best_from_grid}
\alias{select_best_from_grid}
\title{Select the best configuration from a tuning table}
\usage{
select_best_from_grid(
  tuning_df,
  selection_metric = c("euclid", "rmse", "r2", "rpiq")
)
}
\arguments{
\item{tuning_df}{data.frame containing RMSE_val, R2_val, RPIQ_val}

\item{selection_metric}{one of: "euclid", "rmse", "r2", "rpiq"}
}
\value{
list(best=one-row data.frame, best_score=numeric)
}
\description{
Select the best configuration from a tuning table
}
