% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_setup.R
\name{vae_configure}
\alias{vae_configure}
\title{Configure Python / reticulate for soilVAE}
\usage{
vae_configure(python = NULL, venv = NULL, conda = NULL, persist = TRUE)
}
\arguments{
\item{python}{Path to a Python executable.}

\item{venv}{Name/path of a virtualenv.}

\item{conda}{Name of a conda environment.}

\item{persist}{If TRUE, stores the choice in \code{options()} for reuse within the current R session.}
}
\value{
Invisibly TRUE.
}
\description{
Selects a Python environment for use with \'reticulate\'.
This function \strong{does not} install Python packages.
}
