% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vae_predict.R
\name{vae_encode}
\alias{vae_encode}
\title{Extract latent embeddings (z) from a fitted VAEReg model}
\usage{
vae_encode(model, X)
}
\arguments{
\item{model}{A fitted model returned by \code{vae_build()} and trained with
\code{vae_fit()}.}

\item{X}{Matrix-like predictors.}
}
\value{
A numeric matrix of embeddings.
}
\description{
Extract latent embeddings (z) from a fitted VAEReg model
}
