% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R, R/vae_fit.R
\name{vae_fit}
\alias{vae_fit}
\title{Fit VAEReg}
\usage{
vae_fit(
  model,
  X,
  y,
  X_val = NULL,
  y_val = NULL,
  epochs = 80L,
  batch_size = 64L,
  patience = 10L,
  verbose = 0L
)

vae_fit(
  model,
  X,
  y,
  X_val = NULL,
  y_val = NULL,
  epochs = 80L,
  batch_size = 64L,
  patience = 10L,
  verbose = 0L
)
}
\arguments{
\item{model}{Python VAEReg object from vae_build()}

\item{X}{matrix (n x p)}

\item{y}{numeric vector (n)}

\item{X_val}{optional matrix}

\item{y_val}{optional numeric vector}

\item{epochs}{integer}

\item{batch_size}{integer}

\item{patience}{integer for early stopping (only if validation provided)}

\item{verbose}{0/1/2}
}
\value{
training history (Python object)
}
\description{
Fit VAEReg

Fit VAEReg
}
