% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vae_predict.R
\name{vae_predict}
\alias{vae_predict}
\title{Predict using a fitted VAEReg model}
\usage{
vae_predict(model, X)
}
\arguments{
\item{model}{A fitted model returned by \code{vae_build()} and trained with
\code{vae_fit()}.}

\item{X}{Matrix-like predictors.}
}
\value{
Numeric vector of predictions.
}
\description{
Predict using a fitted VAEReg model
}
